utils/pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "openai>=1.0.0",
    "tqdm>=4.66.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.1",
    "asyncpg>=0.30.0",
]
